docker files for datanode/master

+++ datanode/test_data.py
@@ -15,8 +15,11 @@ def client():
 
 
 def init_storage_test():
-    rmtree(FILE_STORE)
-    os.mkdir(FILE_STORE)
+    if os.path.exists(FILE_STORE):
+        rmtree(FILE_STORE)
+        os.mkdir(FILE_STORE)
+    else:
+        os.mkdir(FILE_STORE)
 
 
 def test_filesystem_delete(client):
